Remove UserProfile model
Turns out Django's built‐in User model already has name fields.
Since that was all we were adding,
the extra model is unnecessary.

The fact that it assumes First‐ and Last‐Name stucture
is a bug (albeit a common one),
but we can stick to just using the former if we're only in need of one name
as per spec.