Pull out CSS into separate file
About time this was done.

The Edit Habit page seems to have changed slightly
(I suppose we could fix it back if necessary)
but seems to be relatively littl damage.

The CSS file itself could do with a lot of cleaning up
(I just dumped all the CSS from the individual files
one after the other)
but that's a separate thing.